:hammer: refactor: MenuScheduler 파일구조, 이름 변경
