+++ store-service/src/main/java/com/example/cloudfour/storeservice/domain/menu/scheduler/MenuMongoScheduler.java
@@ -1,26 +1,19 @@
-package com.example.cloudfour.storeservice.scheduler;
+package com.example.cloudfour.storeservice.domain.menu.scheduler;
 
-import com.example.cloudfour.storeservice.domain.collection.converter.DocumentConverter;
-import com.example.cloudfour.storeservice.domain.collection.document.ReviewDocument;
-import com.example.cloudfour.storeservice.domain.collection.document.StoreDocument;
 import com.example.cloudfour.storeservice.domain.collection.repository.command.MenuCommandRepository;
-import com.example.cloudfour.storeservice.domain.collection.repository.command.ReviewCommandRepository;
-import com.example.cloudfour.storeservice.domain.collection.repository.command.StoreCommandRepository;
+import com.example.cloudfour.storeservice.domain.collection.repository.command.StockCommandRepository;
 import com.example.cloudfour.storeservice.domain.common.enums.SyncStatus;
 import com.example.cloudfour.storeservice.domain.menu.entity.Menu;
 import com.example.cloudfour.storeservice.domain.menu.entity.MenuOption;
+import com.example.cloudfour.storeservice.domain.menu.entity.Stock;
 import com.example.cloudfour.storeservice.domain.menu.repository.MenuOptionRepository;
 import com.example.cloudfour.storeservice.domain.menu.repository.MenuRepository;
-import com.example.cloudfour.storeservice.domain.review.entity.Review;
-import com.example.cloudfour.storeservice.domain.review.repository.ReviewRepository;
-import com.example.cloudfour.storeservice.domain.store.entity.Store;
-import com.example.cloudfour.storeservice.domain.store.repository.StoreRepository;
+import com.example.cloudfour.storeservice.domain.menu.repository.StockRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -31,47 +24,15 @@ import java.util.stream.Collectors;
 @Slf4j
 @Transactional
 @RequiredArgsConstructor
-public class MongoCreatedSyncScheduler {
-    private final ReviewCommandRepository reviewCommandRepository;
-    private final ReviewRepository reviewRepository;
-    private final StoreCommandRepository storeCommandRepository;
-    private final StoreRepository storeRepository;
-    private final MenuCommandRepository menuCommandRepository;
+public class MenuMongoScheduler {
+
     private final MenuRepository menuRepository;
     private final MenuOptionRepository menuOptionRepository;
+    private final MenuCommandRepository menuCommandRepository;
+    private final StockRepository stockRepository;
+    private final StockCommandRepository stockCommandRepository;
 
-    //@Scheduled(cron = "0 0 0 * * *")
-    @Scheduled(cron = "0 * * * * *")
-    public void createStore(){
-        log.info("MongoDB에 가게 동기화 시작");
-        List<Store> stores = storeRepository.findAllBySyncStatus(SyncStatus.CREATED_PENDING);
-        if(stores.isEmpty()){
-            log.info("MongoDB에 생성할 가게가 존재하지 않음");
-            return;
-        }
-        List<StoreDocument> storeDocuments = stores.stream().map(DocumentConverter::toStoreDocument).toList();
-        storeCommandRepository.saveAll(storeDocuments);
-        stores.forEach(Store::syncCreated);
-        storeRepository.saveAll(stores);
-        log.info("MongoDB에 가게 동기화 완료");
-    }
-
-    @Scheduled(cron = "0 * * * * *")
-    public void createReview(){
-        log.info("MongoDB에 리뷰 동기화 시작");
-        List<Review> reviews = reviewRepository.findAllBySyncStatus(SyncStatus.CREATED_PENDING);
-        if(reviews.isEmpty()){
-            log.info("MongoDB에 생성할 리뷰가 존재하지 않음");
-            return;
-        }
-        List<ReviewDocument> reviewDocuments = reviews.stream().map(DocumentConverter::toReviewDocument).toList();
-        reviewCommandRepository.saveAll(reviewDocuments);
-        reviews.forEach(Review::syncCreated);
-        reviewRepository.saveAll(reviews);
-        log.info("MongoDB에 리뷰 동기화 완료");
-    }
-
-    @Scheduled(cron = "0 * * * * *")
+    @Scheduled(cron = "0 0 4 * * *")
     public void createMenu(){
         log.info("MongoDB에 메뉴 동기화 시작");
         List<Menu> unsyncedMenus = menuRepository.findAllBySyncStatus(SyncStatus.CREATED_PENDING);
@@ -89,7 +50,7 @@ public class MongoCreatedSyncScheduler {
         log.info("MongoDB에 메뉴 동기화 완료");
     }
 
-    @Scheduled(cron = "0 * * * * *")
+    @Scheduled(cron = "0 0 4 * * *")
     public void createMenuOption(){
         log.info("MongoDB에 메뉴옵션 동기화 시작");
         List<MenuOption> unsyncedMenuOptions = menuOptionRepository.findAllBySyncStatus(SyncStatus.CREATED_PENDING);
@@ -105,5 +66,47 @@ public class MongoCreatedSyncScheduler {
         unsyncedMenuOptions.forEach(MenuOption::syncCreated);
         menuOptionRepository.saveAll(unsyncedMenuOptions);
         log.info("MongoDB에 메뉴 옵션 동기화 완료");
+    }
+
+    @Scheduled(cron = "0 * * * * *")
+    public void deleteMenu(){
+        log.info("MongoDB에 메뉴 삭제 동기화 시작");
+        List<Menu> menus = menuRepository.findAllByIsDeleted();
+        if(menus.isEmpty()){
+            log.info("삭제할 메뉴 데이터 없음");
+            return;
+        }
+        List<UUID> menuIds = menus.stream().map(Menu::getId).toList();
+        menuCommandRepository.deleteAllByMenuIdIn(menuIds);
+        log.info("MongoDB에 메뉴 삭제 완료");
+    }
+
+    @Scheduled(cron = "0 0 4 * * *")
+    public void deleteMenuOption(){
+        log.info("MongoDB에 메뉴 옵션 삭제 동기화 시작");
+        List<MenuOption> menuOptions = menuOptionRepository.findAllByIsDeleted();
+        if(menuOptions.isEmpty()){
+            log.info("삭제할 메뉴 옵션 데이터 없음");
+            return;
+        }
+        List<UUID> menuOptionIds = menuOptions.stream().map(MenuOption::getId).toList();
+        menuCommandRepository.deleteAllByMenuOptionIdIn(menuOptionIds);
+        log.info("MongoDB에 메뉴 옵션 삭제 완료");
+    }
+
+    @Scheduled(cron = "0 0 4 * * *")
+    public void refreshQuantity(){
+        log.info("MongoDB에 수량 최신화 시작");
+        List<Stock> pendingStocks = stockRepository.findAllBySyncStatus(SyncStatus.UPDATED_PENDING);
+        if(pendingStocks.isEmpty()){
+            log.info("MongoDB에 최신화할 수량아 존재하지 않음");
+        }
+
+        for(Stock stock: pendingStocks){
+            stockCommandRepository.updateStockByMenuId(stock.getMenu().getId(), stock.getQuantity());
+            stock.setSyncStatus(SyncStatus.UPDATED_SYNCED);
+            stockRepository.save(stock);
+        }
+        log.info("MongoDB에 수량 최신화 완료");
     }
 }
